fix(events): distrust same-app activation re-reports (#463 QA)
Raising window B activates its app; a lazy-AX app (Electron)
can re-report its OLD focused window A — same pid, different
window, on a hidden space — before the raise lands internally.
The cmd-tab focus-follow trusted that report and yanked the
user to A's space right after they explicitly switched away
(#465 QA: space 3 Antigravity → empty space 2 → space 5's
Antigravity jumped back to 3; the new empty-space Finder yield
made the next switch a real app activation, exposing the hole).

While our raise of a same-app window is unechoed
(`outstandingSelfRaises`), a focus report for a sibling window
on a NON-visible space is now treated as the activation echo it
is: state focus reverts, no focus-follow. Visible-space sibling
reports (real clicks) and other-app reports are untouched; a
genuine cmd-tab to the hidden sibling inside the narrow
unechoed window is the accepted trade.

The `.windowFocused` case moves to KiwiCore+FocusEvents.swift
(350-line ceiling). Three new tests pin suppress / clean-follow
/ other-app-no-suppress.

Note: SpaceBarDropCoordinatorTests dwell test flaked once under
full-suite load (330 s run, heavy machine load); passes twice
isolated — #344 environmental pattern, not this change.
