+++ Sources/KiwiDeskCore/App/KiwiCore+Events.swift
@@ -43,157 +43,7 @@ extension KiwiCore {
             handleMonitorChange()
             emitMonitorChange()
         case .windowFocused(let id):
-            // Dismissing an ignored panel (Ghostty's quick
-            // terminal) leaves the app frontmost and makes AX
-            // re-report its managed main window — which may live
-            // on another space — as focused. Following that
-            // stale report switches spaces, or in a focus-driven
-            // layout yanks focus to the main window on the
-            // active space. The panel was flagged active while
-            // it held focus (#21); consume the flag here and
-            // restore the pre-panel focus instead of acting on
-            // the report (#244). A focus landing on any other
-            // app means the panel's app resigned frontmost, so
-            // drop stale flags to keep a later genuine focus of
-            // that app from being suppressed.
-            if let pid = state.windows[id]?.pid {
-                if ignoredPanelActive.remove(pid) != nil {
-                    // The report is spurious, but the id could
-                    // still be an outstanding self-raise; drop
-                    // it so a later genuine focus of this window
-                    // is not misread as our own echo (cf. the
-                    // .windowDestroyed cleanup below).
-                    outstandingSelfRaises.remove(id)
-                    if let before = effects.focusBefore,
-                        let space = state.workspaces.space(
-                            of: before
-                        )
-                    {
-                        state.workspaces.focus(
-                            before,
-                            in: space
-                        )
-                    }
-                    return
-                }
-                ignoredPanelActive.removeAll()
-            }
-            // A focus echo from our own z-order raise (#418/#425):
-            // AX couples the raise with app activation, so raising a
-            // float above the tiled plane, or a pile member during a
-            // cascade restore, emits a focus echo for it. The raise
-            // must not move focus or the active ring onto that window
-            // — revert to the window the user actually reached and
-            // drop the echo, leaving it raised. The stamped set never
-            // holds the intended focus (the raisers skip it), so a
-            // deliberate focus (a space switch onto a float; a click
-            // on a pile-mate) is unstamped and falls through.
-            // Never consumed for an echo of our OWN focus raise
-            // (`outstandingSelfRaises`): a keyboard focus can land
-            // on a window a pile restore stamped moments earlier,
-            // and for a tiled-sticky traveler `focusBefore` stays
-            // on the stale local slot (the membership guard keeps
-            // `space.focused` off travelers), so `intended != id`
-            // cannot clear it — the revert would undo a deliberate
-            // focus and strand the ring off the truly focused
-            // window (#431). The self-echo path below consumes the
-            // outstanding entry; the stamp expires on its own.
-            if let stamp = zOrderRaiseEchoes[id],
-                Date().timeIntervalSince(stamp)
-                    < Self.zOrderRaiseEchoWindow,
-                let intended = effects.focusBefore, intended != id,
-                !outstandingSelfRaises.contains(id)
-            {
-                zOrderRaiseEchoes[id] = nil
-                if let space = state.workspaces.space(of: intended) {
-                    state.workspaces.focus(intended, in: space)
-                }
-                updateBorders()
-                updateStickyIndicators()
-                return
-            }
-            // Echo provenance (#152/#158): an echo of KiwiDesk's
-            // own AX raise is not a user action. When one lands
-            // after focus has already moved on in the active
-            // scrolling space — a later raise, deferred or
-            // forward-immediate — the echo (and the state focus
-            // StateCoordinator just moved onto the echoed window)
-            // would revert to the stale target. Re-assert the
-            // intended focus and drop the echo. Consume the id
-            // from the outstanding set either way.
-            let selfEcho = outstandingSelfRaises.remove(id) != nil
-            if selfEcho,
-                activeSpace?.mode.defersFocusRaise == true,
-                let intended = effects.focusBefore, intended != id,
-                state.workspaces.space(of: id)
-                    == state.workspaces.activeSpace
-            {
-                if let space = state.workspaces.space(
-                    of: intended
-                ) {
-                    state.workspaces.focus(intended, in: space)
-                }
-                return
-            }
-            // A real focus echo (a user click mid-pan) or a self
-            // echo that matches the intended focus supersedes the
-            // deferred raise (#143): the OS already raised whatever
-            // the user reached, and a stale raise firing after the
-            // pan would steal focus back.
-            pendingFocusRaise = nil
-            emitFocusChange(id)
-            // Move the focus ring to the newly focused window.
-            // Static layouts don't retile on focus (below), so
-            // this is the only refresh they get; focus-driven
-            // layouts retile and refresh again (cheap, idempotent).
-            updateBorders()
-            // Chips too: the click's raise put the window ABOVE
-            // its own chip, and chip stacking is asserted only
-            // on sync (never per tick) — without this the mark
-            // vanished behind the window on a plain click until
-            // the next retile (owner QA 2026-07-21).
-            updateStickyIndicators()
-            // The Space Bar's focused-glyph accent follows the
-            // same rule (#293): layout-independent, so it can't
-            // ride the focus-driven retile alone.
-            updateSpaceBar()
-            // Warp only for focus changes KiwiDesk did not
-            // make itself (cmd+tab, app-driven focus): a
-            // self-raise already warped at intent time in
-            // `focusWindow`, and an echo that moved no focus
-            // (same-window re-focus) is not a change (#186).
-            if !selfEcho, effects.focusBefore != id {
-                warpMouseToFocused(id)
-            }
-            // cmd+tab (or a click) can reach a window hidden
-            // in an inactive virtual space; pull that space
-            // forward instead of typing into a stashed window.
-            // Deferred: app activation transiently re-reports
-            // the app's OLD focused window right before a new
-            // window opens — only follow if focus settles.
-            // A STICKY window is exempt (#414): it is visible
-            // and legitimately focusable on every space, so
-            // following its home would yank the user back to
-            // it on every interaction — the fly-back bug.
-            if let space = state.workspaces.space(of: id),
-                space != state.workspaces.activeSpace,
-                state.windows[id]?.isSticky != true
-            {
-                scheduleFocusFollow(id)
-            } else if activeSpace?.mode.isFocusDriven == true {
-                // A foreign STICKY focus also lands here (its
-                // follow is exempt above): the retile is
-                // idempotent — the fold set focus in the
-                // window's home space, not this one — so it
-                // costs one no-op pass. Deliberate.
-                retileWithScrollDuration()
-            }
-            // Keep floats above the just-focused tiled window
-            // (#418). Skipped on a self-echo so the raise's own
-            // focus handoff cannot re-trigger it.
-            if !selfEcho {
-                raiseFloatsAbove(afterFocusing: id)
-            }
+            handleWindowFocused(id, effects: effects)
         case .windowCreated(let window):
             // A brand-new window supersedes a pending follow
             // of a hidden window (see above).
